# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rocket-api",
    "description": "API Documentation for Rocket API",
    "version": "1.0.0"
  },
  "host": "myrocketapp-34740598b7f1.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/rockets/": {
      "get": {
        "tags": [
          "Rockets"
        ],
        "summary": "Get all rockets",
        "description": "Routes for managing rockets",
        "parameters": [
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved rockets",
            "schema": {
              "$ref": "#/definitions/Rocket"
            }
          },
          "400": {
            "description": "Error fetching rockets",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Rockets"
        ],
        "summary": "Create a new rocket",
        "description": "Routes for managing rockets",
        "parameters": [
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Falcon 9"
                },
                "agencyId": {
                  "type": "string",
                  "example": "5e79dfb4a61a9c001722e492"
                },
                "payloadCapacity": {
                  "type": "number",
                  "example": 22800
                },
                "fuelType": {
                  "type": "string",
                  "example": "liquid"
                },
                "launchDate": {
                  "type": "string",
                  "example": "2024-05-01"
                }
              },
              "required": [
                "name",
                "agencyId",
                "payloadCapacity",
                "fuelType",
                "launchDate"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created rocket",
            "schema": {
              "$ref": "#/definitions/Rocket"
            }
          },
          "400": {
            "description": "Invalid request body",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/rockets/{id}": {
      "get": {
        "tags": [
          "Rockets"
        ],
        "summary": "Get a rocket by ID",
        "description": "Routes for managing rockets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Rocket ID"
          },
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved rocket",
            "schema": {
              "$ref": "#/definitions/Rocket"
            }
          },
          "400": {
            "description": "Invalid rocket ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Rocket not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Rockets"
        ],
        "summary": "Update a rocket by ID",
        "description": "Routes for managing rockets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Rocket ID"
          },
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "agencyId": {
                  "example": "any"
                },
                "payloadCapacity": {
                  "example": "any"
                },
                "fuelType": {
                  "example": "any"
                },
                "launchDate": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated rocket",
            "schema": {
              "$ref": "#/definitions/Rocket"
            }
          },
          "400": {
            "description": "Invalid rocket ID or request body",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Rocket not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Rockets"
        ],
        "summary": "Delete a rocket by ID",
        "description": "Routes for managing rockets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Rocket ID"
          },
          {
            "name": "x-auth-token",
            "description": "Authorization token",
            "required": true,
            "type": "string",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted rocket",
            "schema": {
              "$ref": "#/definitions/Rocket"
            }
          },
          "400": {
            "description": "Invalid rocket ID",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Rocket not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "example": "string"
        }
      }
    },
...
    "Rocket": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "string"
        },
        "agencyId": {
          "type": "string",
          "example": "ObjectId"
        },
        "payloadCapacity": {
          "type": "string",
          "example": "integer"
        },
        "fuelType": {
          "type": "string",
          "example": "string"
        },
        "launchDate": {
          "type": "string",
          "example": "date"
        }
      }
    }
  },
...
}